Convert youki crate to a library
